Extracted the resilience strategies' methods into a separate service.
