+++ Services/Infrastructure/OJS.Services.Infrastructure/ResilienceStrategies/Enums/ResilienceStrategyEventType.cs
@@ -1,0 +1,11 @@
+﻿namespace OJS.Services.Infrastructure.ResilienceStrategies.Enums;
+
+public enum ResilienceStrategyEventType
+{
+    OnCircuitOpened,
+    OnCircuitClosed,
+    OnCircuitHalfOpened,
+    PipelineExecuting,
+    PipelineExecuted,
+    OnRetry,
+}
